Finishing up an epic refactor.  This bears some explanation.
So, I discovered that I'd misunderstood how useEffect's cleanup()
functions work - when they fire and what data they fire with.

I discovered this, when I realized that I wasn't storing the requestId
on the RequestTracker objects like I thought I was, so that when I was
passing them to various cleanupRequest() methods, nothing was happening.
RequestTrackers were just building up in the redux store and memory was
leaking like crazy.

So I fixed that issue, adding requestId to the RequestTracker, and all
hell broke lose.  Because the timing of the cleanupRequest() calls was
now all wrong, and requests were being cleaned up before we were done
with them.

This sent me down a deep, deep rabbithole for a while, because the
request object will be updated multiple times over its lifecycle and we
don't want to it be cleaned up until either we're done with that request
and on to the next one or the component unmounts.  There was no
configuration of useEffect and cleanup() that I could come up with that
matched that timing.  At least, not with the request as a dependency.

I eventually figured out that the update timing of the requestId
perfectly matched when we'd want to cleanup a request and so I
refactored *every single call to cleanupRequest()* to live in a
useEffect() cleanup() method with the appropriate requestId as a
dependency.

This then introduced a new problem: all kinds of render thrashing and
unnecessary component reupdating.  You see, I hadn't actually been
cleaning up our requests.  So I didn't realize just how often they were
getting remade, and now that they weren't cached indefinitely in the
store, it was causing components to re-render.

This lead me down a new rabbithole of figuring out how to cache requests
for a reasonable amount of time so that components didn't completely
re-render all their data on every single view.

And thus we landed with the refactor you see here.  We're now in a
*MUCH* better place than we were.  But this was a 2 or 3 day side quest
I wasn't expecting to that.  That said, we have a much more centralized
space for request tracking and a cache system built into it that will
give us much more control over when and how we make requests and how
much we reuse the data they return.  This will help us manage load on
the backend and run more cheaply.

Plus, I have a much better understanding of React's actual lifecycle (I
think... I hope...) than I did, and I did a fuck ton of cleanup in the
process of refactoring.  So the alpha will (should... I hope...) be much
more stable now than it would have been before.