Text overlay is working on the published paper page.
Had to use ReactPDF v6.0.0-beta.3 to get it, but hey, I'll take it.
We can look at converting the review page to ReactPDF later, that will
allow us to enable text overlay and probably save us a ton in the
bundle, because we won't be loading PDF.js twice!