Issue #48 - Review UX with no overlapping comments.
After three days of fighting with redux and react I finally have a
review UX that works and isn't awful.  I'm bummed I couldn't get Google
Docs style comments working - though... I might be able to now
actually... but what I have now is honestly probably better in the long
run.  It'll be much easier to make this responsive and to make it work
on small screens.  So yeah, we're going with it.