Issue #68 - Paper Lists report accurate information.
Also rolled back our ReactPDF version off of 6.0.0-beta.3, meaning we
lose working text overlay (for now).  There was a bug we encoutered
where it would hang loading for ever when the source PDF changed - which
happens when you switch version on the review page.  It's a bummer, but
it is what it is.  Hopefully they fix it before we go to beta.