Database optimizations and a small bug fixes in responses.
We hadn't been thinking too hard about where we might need indexes when
writing the schema.  But after noticing the paper list queries taking
nearly half a second on stage, figured it was worth doing a naive
optimization pass on the database and adding in any missing indexes on
foriegn key constraints.  Sure enough, cut the paper list query repsonse
time in half.

In the process, we discovered a bug in responses, where
response_versions were having version set to null.  Fixed it.