Issue #96 - Reviews in progress can now be cancelled.
Took on some tech debt with this one.  There's a race condition where,
if you go straight from editting a comment in progress to cancelling a
review the patch and delete requests can interleave and cause the
deleted review to reappear post deletion.  I handled it by looking for
delete requests in progress from the patch request and bailing out if
they are found.  There's got to be a better way, but it works for now.

It logs an error when it happens so we can track how often it happens.
It really should be a warning, so we should differentiate warnings from
errors at some point.  I'll leave that for the error handling story.