Issue #34 - Get Initial Reputation from Citations. WIP.
This one's a doozy.  In the course of preparing to initialize reputation
from citations read from OpenAlex, I discovered OpenAlex already has
"concepts" which are a close analog for Peer Review's fields.  They ave
65000 of them, already linked to authors and works and citation counts.

It's going to make initializing the reputation system a piece of cake,
but even better it gave me a *very* good head start on the field system.
I pulled in their concepts and translated them to Peer Review fields.
That took some doing, OpenAlex is new and there was some bad data.

Then I had to do some work to optimize staring up the local database
container, because loading 65000 fields and over 100,000 relationships
between them was taking a long time.  I managed to get it down to a few
minutes from bare SQL initialization, and a few seconds from a dump
file.

Then I discovered that the field code I had written (intentionally naive
and unoptimized) could not handle 65000 fields.  So I went down the
rabbithole of some base level optimizations to get it to a place where
it could handle the new scale of the field database.

Finally I got to work on reputation initialization and here we are.
Realized I hadn't pushed all week (becuase it constantly felt like it
was not in a pushable state through this work).  So I finally took a
pause and got things in a pushable state.

More to do on reputation initalization, but I think I'm close.