Issue #34 + Issue #86 - Major refactor and working towards auto-generating reputation from ORCID iD.
Was working towards auto-generating reputation from ORCID iD when I
finally came up with a way to reduce the front-end request boilerplate.
I've taken a shot at this refactor a couple of times and backed it out
each time when I discovered an issue with the approach I'd taken.  I
just hadn't come up with a satisfactory way to unify making the request
and handling the errors.  But as the boilerplate has grown, the need for
unifying it has grown as well.  I finally came up with an interface I
don't hate and which seems to cover all the cases, so I'm running with
it!