Issue #86 - Maybe fixing an error in the tests?
The tests broke in the pipeline on an error that doesn't appear locally,
related to FormData not being available in the node environment.  For
whatever reason, jest isn't throwing that error when the test run
locally.  But it is when run in Github Actions.

Installing this node module should (in theory) fix it.