Adding a model.  It is time.
Made the initial decision not to create a model in order to minimize the
necessary Object Relational Mapping work.  We ended up doing a fair
amount of that in the DAOs anyway.  We were able to reasonably maintain
consistency across the backend, frontend, and tests, but it's starting
to become difficult. And with the pivot to JournalHub, the complexity
has increased to where we now need a much clearer source of truth for
the structure of the data and what exactly we're doing with all of it.

So it's time to add a model.  This is a model that will be used on both
the frontend and backend so we're defining it in its own package so that
it can be imported everywhere.